requirements.txt
jdatetime
djangorestframework
asgiref==3.3.4
Django==3.2
django-debug-toolbar==3.2.1
django-filter==2.4.0
future==0.18.2
Markdown==3.3.4
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
